refactor(system-users): os pkg carved out of host pkg #672
