+++ k2s/cmd/k2s/cmd/common/common.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils/logging"
 	"github.com/siemens-healthineers/k2s/internal/core/config"
-	"github.com/siemens-healthineers/k2s/internal/host"
 	bl "github.com/siemens-healthineers/k2s/internal/logging"
+	"github.com/siemens-healthineers/k2s/internal/os"
 	"github.com/siemens-healthineers/k2s/internal/powershell"
 
 	"github.com/siemens-healthineers/k2s/internal/core/setupinfo"
@@ -99,7 +99,7 @@ func NewPtermWriter() *PtermWriter {
 	}
 }
 
-func NewSlogWriter() host.StdWriter { return &SlogWriter{} }
+func NewSlogWriter() os.StdWriter { return &SlogWriter{} }
 
 func NewCmdContext(config *config.Config, logger *logging.Slogger) *CmdContext {
 	return &CmdContext{
