+++ k2s/internal/containernetworking/network.go
@@ -16,7 +16,7 @@ import (
 	"github.com/Microsoft/windows-container-networking/cni"
 	"github.com/Microsoft/windows-container-networking/common"
 	"github.com/Microsoft/windows-container-networking/network"
-	"github.com/siemens-healthineers/k2s/internal/host"
+	kos "github.com/siemens-healthineers/k2s/internal/os"
 	"github.com/sirupsen/logrus"
 
 	"github.com/Microsoft/hcsshim/hcn"
@@ -277,7 +277,7 @@ func (plugin *netPlugin) Add(args *cniSkel.CmdArgs) (resultError error) {
 				logrus.Debugf("[cni-net] XXXX Current working directory: %s", path)
 			}
 
-			pathExe, errPath := host.ExecutableDir()
+			pathExe, errPath := kos.ExecutableDir()
 			if errPath != nil {
 				logrus.Debugf("[cni-net] XXXX Current directory Error:", errPath)
 				return
