+++ k2s/test/e2e/cli/cmd/stop/nosetup/corruptedstate/stop_corrupted-state_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/internal/core/config"
 	"github.com/siemens-healthineers/k2s/internal/core/setupinfo"
-	"github.com/siemens-healthineers/k2s/internal/host"
+	kos "github.com/siemens-healthineers/k2s/internal/os"
 
 	"testing"
 
@@ -53,7 +53,7 @@ var _ = Describe("stop", Ordered, func() {
 		inputData, err := json.Marshal(inputConfig)
 		Expect(err).ToNot(HaveOccurred())
 
-		currentDir, err := host.ExecutableDir()
+		currentDir, err := kos.ExecutableDir()
 		Expect(err).ToNot(HaveOccurred())
 		installDir := filepath.Join(currentDir, "..\\..\\..\\..\\..\\..\\..\\..")
